refactor: 복잡한 생성 로직 factory 패턴 적용
