+++ src/main/java/com/rudkids/rudkids/domain/item/ItemFactory.java
@@ -1,0 +1,7 @@
+package com.rudkids.rudkids.domain.item;
+
+import com.rudkids.rudkids.domain.item.domain.Item;
+
+public interface ItemFactory {
+    Item create(ItemCommand.CreateItemRequest command);
+}
