+++ src/main/java/com/rudkids/rudkids/domain/product/ProductFactory.java
@@ -1,0 +1,7 @@
+package com.rudkids.rudkids.domain.product;
+
+import com.rudkids.rudkids.domain.product.domain.Product;
+
+public interface ProductFactory {
+    Product create(ProductCommand.CreateRequest command);
+}
